make private key combination / subtraction work
The scalar was sill in the keygen's privkey field when requesting
the private key from the underlying wallet.  This swaps it out and
clears it before requesting the base private key.

Still kindof ugly because it requires a ~superfluous call to
get a private key and operate on it.  but works and pretty concise.