Change lit-af to have one fee command
We don't have the issue of missing json keys being set to
zero value in lit-af (we get the command before it goes to
json) so we can use one command in a non-ugly way.

Entering "fee" will query the current fee.
Entering "fee 200" will set the fee to 200 sat / byte.
That seems pretty clean.