Re-register refund addresses with wallit after restarting lit
When a channel is first funded your refund address is registered with wallit so the resulting UTXO is picked up by a close. After Lit was restarted we were not re-registering those refund addresses with the wallit. As a result, closing a channel resulted in you receiving no new UTXO.