Fix race condition in channel push code (#167)
* Fix bug where non-witness outputs were selected in PickUtxos

If there were a mixture of witness and non-witness outputs in the db then PickUtxos would sometimes select non-witness inputs even when only-witness was set. This prevented channel funding in a number of cases.

* Don't pass LitNode by-value to message handlers

In Go, passing a mutex by value copies it and creates a new mutex. Thus if you pass a containing struct by-value and lock the mutex, you're locking a copy rather than the original

* Don't build the justice tx in a coroutine, could be a race condition

* Missed merge conflicts

* Use consts.MinOutput