Add ipv6 to lndc and uspv (#201)
* Handle updates to the tracker for IPv6 support

Sometimes our default interface is IPv6, and Lit currently cannot connect to IPv6 nodes. Therefore I updated lit-tracker to accept both an IPv4 and IPv6 address on announce. This update sends both types of IP address to the tracker if available, but only attempts to connect to the IPv4 version for now. LNDC needs to be updated to handle IPv6 to use the other address potentially returned by the tracker.

* Add IPv6 support to uspv and lndc

* Correct error

* Rebase upon master and #200