Made DialNode return the connection instead of magically setting it on ourselves.
In the future we may want to pull it out of the SPVCon type into its own thing, and pass it a list of string to try to connect to.

Fixed a bunch of Go linter error/warnings.

Still some left, but those ones are stupid and we should ignore them.

Applied suggestions.

Reapplied some changed that got reverted somewhere in connection.

Detect whether coin daemon is running before funding (#255)

Preliminary NAT traversal (#220)

* Add UPnP port forwarding

* Add PmP and move UPnP functions to separate lib

* Punch NAT automatically while listening

add notes on invoicing / prompt

fix triple backtick problem

RPC improvements (merged Accept/Decline, detail on Fund, plain HTTP endpoint, etc.) (#279)

* Made specific FundReply message with some more specific information.

* Merged some RPC calls into just DualFundRespond and ContractRespond.

* Added a plain HTTP RPC endpoint, plus a debug log RPC call.

* Moved litrpc API doc, pls make this merge werk.

Added some delays in the integration test framework.

Fixed not setting connection object after dialing.

New ncurses UI for monitoring Lit a la htop (#295)

* Set up project for lit-af in Rust.

* Added root Cargo.toml to make the Atom RLS plugin happy.

* Added basic JSON-RPC connection for the Lit node.

* Converted the Rust litrpc impl to use the one-off HTTP endpoint.

* Use more correct error checking for RPC calls.

* Added macros to generate most of the core LitRPC commands as normal functions.

* Updated gitignore.

* Set up basic ncurses TUI.

* Refactored UI generation code to make it update from the Lit daemon.

* Added BSOD.

* Added views for balances, addresses, etc.  Had to add a new RPC command for it.

* Moved address list into balances, plus minor cleanup.

* Lots of changes to support adding a "last update" timestamp to channels incl in the UI.

* s/lit-af-rs/rs-af/g

* Fix Cargo.

* Fixed unset LastUpdate field in RPC ChannelInfo struct.

* Added run script, plus some minor cleanup.

* unrelated: Don't delete the webui in the Makefile anymore

* Update litrpc doc for GetAddresses call.

* Fixed unused import.

Fix segfault caused by nil htlcTxo (#302)

When searching for the desired htlcTxo in a commitment transaction, if it wasn't found it returned nil. Instead return an error and check for it to stop the nil pointer dereference

Made the Lit-AF ls command only print the things you tell it to. (#270)

* Cleaned up control flow of the DialNode function.

* Made DialNode return the connection instead of magically setting it on ourselves.

In the future we may want to pull it out of the SPVCon type into its own thing, and pass it a list of string to try to connect to.

* Fixed a bunch of Go linter error/warnings.

Still some left, but those ones are stupid and we should ignore them.

* Applied suggestions.

* Reapplied some changed that got reverted somewhere in connection.

* Started better printout for ls command.

* Added actual new argument checking to ls command.

* Misc cleanup and making things more consistent.

* Reordered ls no-argument help text.

* Tweaked when headings are displayed in `ls -a`.

* Fixed duplicate declaration.

* Fixed issue with uspv backend node assignment.